+++ ast/parser.go
@@ -2,479 +2,633 @@ package ast
 
 import (
 	"errors"
-	"fmt"
 )
 
-const (
-	block_none = iota
-	h1
-	h2
-	h3
-	h4
-	h5
-	h6
-	ul
-	p
-	code_block
+type stateFunc func(s *parseState, char byte) (stateFunc, error)
 
-	stateReadBlock = iota
-	read_head_space
+type parseState struct {
+	src          string
+	index        int
+	srcLen       int
+	root         *Block
+	currentBlock *Block
+	blockStack   *blockStack
 
-	read_bq_1
-	read_bq_2
-	read_bq_3
-	code
-	read_bq_end_1
-	read_bq_end_2
-	read_bq_end_3
+	textValue      []byte
+	linkTitleValue []byte
+	linkURLValue   []byte
+	attrName       []byte
+	attrValue      []byte
 
-	read_ul
-	read_h1
-	read_h2
-	text
-	textLinkStart
-	textLinkEnd
-	textLinkUrlStart
-	text_link_url_end
-	textImage1
-	textImageStart
-	textImageEnd
-	textImageURLStart
-	textImageURLAttrKey
-	textImageURLAttrValue
-	text_image_url_end
-	text_br
-	textCodeStart
-	textCodeEnd
-	textInlineCode
-)
+	hCount int
+}
 
-func Parse_(src string) (*Block, error) {
-	var textValue []byte
-	var attrKey []byte
-	var attrValue []byte
-	state := stateReadBlock
-	index := 0
-	srcLen := len(src)
+// Parse src markdown to block
+func Parse(src string) (*Block, error) {
 	root := newBlock(TypeRoot)
-	currentBlock := root
-	blockStack := &blockStack{values: make([]*Block, 0)}
-
-	for index < srcLen {
-		char := src[index]
-		switch state {
-		case stateReadBlock:
-			if char == '#' {
-				state = read_h1
-			} else if char == '`' {
-				state = read_bq_1
-			} else if char == '\n' {
-				if currentBlock.Type == TypeLI {
-					// pop x2
-					blockStack.Pop()
-					blockStack.Pop()
-				}
-				// ignore
-			} else if char == ' ' {
-				state = read_head_space
-			} else {
-				// Add p-text block
-				pBlock := newBlock(TypeP)
-				textBlock := newBlock(TypeText)
-				appendChild(currentBlock, pBlock)
-				appendChild(pBlock, textBlock)
-				blockStack.Push(currentBlock)
-				blockStack.Push(pBlock)
-
-				state = text
-				currentBlock = textBlock
-				textValue = make([]byte, 0)
-				index--
-			}
-		case read_bq_1:
-			if char == '`' {
-				state = read_bq_2
-			} else {
-				// textnode + start code block
-				pBlock := newBlock(TypeP)
-				codeBlock := newBlock(TypeCode)
-				appendChild(currentBlock, pBlock)
-				appendChild(pBlock, codeBlock)
-				blockStack.Push(currentBlock)
-				blockStack.Push(pBlock)
-
-				state = textInlineCode
-				currentBlock = codeBlock
-				textValue = make([]byte, 0)
-				index--
-			}
-		case read_bq_2:
-			if char == '`' {
-				state = read_bq_3
-			} else {
-				pBlock := newBlock(TypeP)
-				textBlock := newBlock(TypeText)
-				appendChild(currentBlock, pBlock)
-				appendChild(pBlock, textBlock)
-				blockStack.Push(currentBlock)
-				blockStack.Push(pBlock)
-
-				state = text
-				currentBlock = textBlock
-				textValue = make([]byte, 0)
-				textValue = appendStr(textValue, "``")
-				index--
-			}
-		case read_bq_3:
-			if char == '\n' {
-
-				codeBlock := newBlock(TypePreCode)
-				textBlock := newBlock(TypeText)
-				appendChild(currentBlock, codeBlock)
-				appendChild(codeBlock, textBlock)
-				blockStack.Push(currentBlock)
-				blockStack.Push(codeBlock)
-
-				state = code
-				currentBlock = textBlock
-				textValue = make([]byte, 0)
-			} else {
-				// TODO : check lang
-				pBlock := newBlock(TypeP)
-				textBlock := newBlock(TypeText)
-				appendChild(currentBlock, pBlock)
-				appendChild(pBlock, textBlock)
-				blockStack.Push(currentBlock)
-				blockStack.Push(pBlock)
-
-				state = text
-				currentBlock = textBlock
-				textValue = make([]byte, 0)
-				textValue = appendStr(textValue, "```")
-
-				index--
-			}
-		case code:
-			if char == '`' {
-				state = read_bq_end_1
-			} else if char == '<' {
-				textValue = appendStr(textValue, "&lt;")
-			} else if char == '>' {
-				textValue = appendStr(textValue, "&gt;")
-			} else {
-				textValue = append(textValue, char)
-			}
-		case read_bq_end_1:
-			if char == '`' {
-				state = read_bq_end_2
-			} else {
-				textValue = append(textValue, '`')
-				textValue = append(textValue, char)
-				state = code
-			}
-		case read_bq_end_2:
-			if char == '`' {
-				state = read_bq_end_3
-			} else {
-				textValue = appendStr(textValue, "``")
-				textValue = append(textValue, char)
-				state = code
-			}
-		case read_bq_end_3:
-			if char == '\n' {
-				currentBlock.Value = string(textValue)
-
-				state = stateReadBlock
-				currentBlock = blockStack.Pop()
-				currentBlock = blockStack.Pop()
-			} else if char == ' ' {
-				// ignore
-			} else {
-				return nil, errors.New(fmt.Sprintf("\\n expected byte %d at %d", char, index))
-			}
-		case read_head_space:
-			if char == '*' || char == '-' {
-				state = read_ul
-			} else {
-				state = text
-			}
-		case read_ul:
-			if char == ' ' {
-				if currentBlock.Type == TypeUL {
-					liBlock := newBlock(TypeLI)
-					textBlock := newBlock(TypeText)
-					appendChild(currentBlock, liBlock)
-					appendChild(liBlock, textBlock)
-					blockStack.Push(currentBlock)
-					blockStack.Push(liBlock)
-
-					state = text
-					currentBlock = textBlock
-					textValue = make([]byte, 0)
-				} else {
-					ulBlock := newBlock(TypeUL)
-					liBlock := newBlock(TypeLI)
-					textBlock := newBlock(TypeText)
-					appendChild(currentBlock, ulBlock)
-					appendChild(ulBlock, liBlock)
-					appendChild(liBlock, textBlock)
-					blockStack.Push(currentBlock)
-					blockStack.Push(ulBlock)
-					blockStack.Push(liBlock)
-
-					state = text
-					currentBlock = textBlock
-					textValue = make([]byte, 0)
-				}
-			} else {
-				return nil, errors.New(fmt.Sprintf("unexpected token at %d", index))
-			}
-		case text:
-			if char == '\n' {
-				parentBlock := blockStack.Top()
-				if parentBlock.Type == TypeLI {
-					currentBlock.Value = string(textValue)
-
-					state = stateReadBlock
-					// pop x 2
-					currentBlock = blockStack.Pop()
-					currentBlock = blockStack.Pop()
-				} else {
-					state = text_br
-				}
-			} else if char == '[' {
-				currentBlock.Value = string(textValue)
-
-				// add link node
-				aBlock := newBlock(TypeAnchor)
-				appendChild(blockStack.Top(), aBlock)
-
-				state = textLinkStart
-				currentBlock = aBlock
-				textValue = make([]byte, 0)
-			} else if char == '!' {
-				state = textImage1
-			} else if char == '`' {
-				currentBlock.Value = string(textValue)
-
-				// add inline code node
-				inlineCodeBlock := newBlock(TypeCode)
-				appendChild(blockStack.Top(), inlineCodeBlock)
-
-				state = textInlineCode
-				currentBlock = inlineCodeBlock
-				textValue = make([]byte, 0)
-			} else if char == '<' {
-				// nop
-			} else if char == '>' {
-				// nop
-			} else {
-				textValue = append(textValue, char)
-			}
-		case textLinkStart:
-			if char == ']' {
-				currentBlock.Value = string(textValue)
-
-				state = textLinkEnd
-				textValue = make([]byte, 0)
-			} else {
-				textValue = append(textValue, char)
-			}
-		case textLinkEnd:
-			if char == '(' {
-				state = textLinkUrlStart
-			} else {
-				return nil, fmt.Errorf("expected is ( but %d at %d", char, index)
-			}
-		case textLinkUrlStart:
-			if char == ')' {
-				currentBlock.URL = string(textValue)
-
-				// add next text block
-				textBlock := newBlock(TypeText)
-				appendChild(blockStack.Top(), textBlock)
-
-				state = text
-				currentBlock = textBlock
-				textValue = make([]byte, 0)
-			} else {
-				textValue = append(textValue, char)
-			}
-		case textImage1:
-			if char == '[' {
-				currentBlock.Value = string(textValue)
-
-				imageBlock := newBlock(TypeImage)
-				appendChild(blockStack.Top(), imageBlock)
-
-				state = textImageStart
-				currentBlock = imageBlock
-				textValue = make([]byte, 0)
-			} else {
-				textValue = append(textValue, '!')
-				index--
-				state = text
-			}
-		case textImageStart:
-			if char == ']' {
-				currentBlock.Value = string(textValue)
-
-				state = textImageEnd
-				textValue = make([]byte, 0)
-			} else {
-				textValue = append(textValue, char)
-			}
-		case textImageEnd:
-			if char == '(' {
-				state = textImageURLStart
-			} else {
-				return nil, fmt.Errorf("expected is ( but %d at %d", char, index)
-			}
-		case textImageURLStart:
-			if char == ')' {
-				currentBlock.URL = string(textValue)
-
-				// set next text block
-				textBlock := newBlock(TypeText)
-				appendChild(blockStack.Top(), textBlock)
-
-				state = text
-				currentBlock = textBlock
-				textValue = make([]byte, 0)
-			} else if char == ' ' {
-				currentBlock.URL = string(textValue)
-				currentBlock.Attributes = make(map[string]string)
-
-				state = textImageURLAttrKey
-				attrKey = make([]byte, 0)
-			} else {
-				textValue = append(textValue, char)
-			}
-		case textImageURLAttrKey:
-			if char == '=' {
-				state = textImageURLAttrValue
-				attrValue = make([]byte, 0)
-			} else {
-				attrKey = append(attrKey, char)
-			}
-		case textImageURLAttrValue:
-			if char == ' ' {
-				currentBlock.Attributes[string(attrKey)] = string(attrValue)
-
-				state = textImageURLAttrKey
-				attrKey = make([]byte, 0)
-			} else if char == ')' {
-				currentBlock.Attributes[string(attrKey)] = string(attrValue)
-
-				// set next text block
-				textBlock := newBlock(TypeText)
-				appendChild(blockStack.Top(), textBlock)
-
-				state = text
-				currentBlock = textBlock
-				textValue = make([]byte, 0)
-			} else {
-				attrValue = append(attrValue, char)
-			}
-		case text_br:
-			if char == '\n' {
-				// end of text and parent block
-				currentBlock.Value = string(textValue)
-				// pop x 2
-				currentBlock = blockStack.Pop()
-				currentBlock = blockStack.Pop()
-
-				state = stateReadBlock
-			} else {
-				state = text
-				textValue = append(textValue, ' ')
-				index--
-			}
-		case read_h1: // #
-			if char == ' ' {
-				h1Block := newBlock(TypeH1)
-				textBlock := newBlock(TypeText)
-
-				appendChild(currentBlock, h1Block)
-				appendChild(h1Block, textBlock)
-				blockStack.Push(currentBlock)
-				blockStack.Push(h1Block)
-
-				currentBlock = textBlock
-				textValue = make([]byte, 0)
-
-				state = text
-			} else if char == '#' {
-				state = read_h2
-			}
-		case read_h2: // ##
-			if char == ' ' {
-				h2Block := newBlock(TypeH2)
-				textBlock := newBlock(TypeText)
-				appendChild(currentBlock, h2Block)
-				appendChild(h2Block, textBlock)
-				blockStack.Push(currentBlock)
-				blockStack.Push(h2Block)
-
-				currentBlock = textBlock
-				textValue = make([]byte, 0)
-
-				state = text
-			}
-		case textInlineCode:
-			if char == '`' {
-				currentBlock.Value = string(textValue)
-
-				// set next text block
-				textBlock := newBlock(TypeText)
-				appendChild(blockStack.Top(), textBlock)
-
-				state = text
-				currentBlock = textBlock
-				textValue = make([]byte, 0)
-			} else if char == '<' {
-				textValue = appendStr(textValue, "&lt;")
-			} else if char == '>' {
-				textValue = appendStr(textValue, "&gt;")
-			} else {
-				textValue = append(textValue, char)
-			}
+	s := &parseState{
+		src:          src,
+		index:        0,
+		srcLen:       len(src),
+		root:         root,
+		currentBlock: root,
+		blockStack:   &blockStack{values: make([]*Block, 0)},
+		hCount:       0,
+	}
+	f := stateReadRootBlock
+	panicCounter := 0
+	for s.index < s.srcLen {
+		panicCounter++
+		if panicCounter > s.srcLen*10 {
+			return nil, errors.New("parser may be in infinte loop")
 		}
-		index++
+		char := src[s.index]
+
+		f2, err := f(s, char)
+		if err != nil {
+			return nil, err
+		}
+
+		f = f2
 	}
-	if currentBlock.Type == TypeText {
-		currentBlock.Value = string(textValue)
+	if s.currentBlock.Type == TypeText {
+		s.currentBlock.Value = string(s.textValue)
+	}
+	return s.root, nil
+}
+
+func stateReadRootBlock(s *parseState, char byte) (stateFunc, error) {
+	if char == ' ' || char == '\n' {
+		// skip
+		s.index++
+		return stateReadRootBlock, nil
+	}
+	if char == '#' {
+		s.hCount = 1
+		s.index++
+		return stateReadHn, nil
+	}
+	if char == '*' || char == '-' {
+		s.index++
+		return stateReadUL, nil
+	}
+	if char == '`' {
+		s.hCount = 1
+		s.index++
+		return stateReadBeginPreCode, nil
+	}
+	// paragraph block
+	pBlock := newBlock(TypeP)
+	textBlock := newBlock(TypeText)
+	appendChild(s.currentBlock, pBlock)
+	appendChild(pBlock, textBlock)
+	s.blockStack.Push(s.currentBlock)
+	s.blockStack.Push(pBlock)
+
+	s.currentBlock = textBlock
+	s.textValue = make([]byte, 0)
+	// read this char as a part of text
+	return stateReadText, nil
+}
+
+func stateReadHn(s *parseState, char byte) (stateFunc, error) {
+	if char == '#' {
+		if s.hCount > 2 {
+			return nil, errors.New("Cannot support level 2 header")
+		}
+		s.hCount++
+		s.index++
+		return stateReadHn, nil
+	}
+	hBlock := newBlock(toHnType(s.hCount))
+	textBlock := newBlock(TypeText)
+
+	appendChild(s.currentBlock, hBlock)
+	appendChild(hBlock, textBlock)
+	s.blockStack.Push(s.currentBlock)
+	s.blockStack.Push(hBlock)
+	s.currentBlock = textBlock
+
+	s.textValue = make([]byte, 0)
+
+	if char == ' ' {
+		return stateFindFirstText, nil
 	}
 
-	return root, nil
+	return stateReadText, nil
 }
 
-func appendChild(b *Block, c *Block) {
-	b.Children = append(b.Children, c)
+func stateFindFirstText(s *parseState, char byte) (stateFunc, error) {
+	if char == ' ' {
+		s.index++
+		return stateFindFirstText, nil
+	}
+	return stateReadText, nil
 }
 
-func appendStr(out []byte, text string) []byte {
-	return append(out, text...)
+func stateReadText(s *parseState, char byte) (stateFunc, error) {
+	if char == '\n' {
+		parentBlock := s.blockStack.Top()
+		if parentBlock.Type == TypeH1 ||
+			parentBlock.Type == TypeH2 {
+			s.currentBlock.Value = string(s.textValue)
+			parentBlock = s.blockStack.Pop() // h block is ended
+			parentBlock = s.blockStack.Pop() // parent of h block
+			s.currentBlock = parentBlock
+
+			s.index++
+			return stateReadRootBlock, nil
+		}
+		if parentBlock.Type == TypeLI {
+			s.currentBlock.Value = string(s.textValue)
+			parentBlock = s.blockStack.Pop() // li block is ended
+			parentBlock = s.blockStack.Pop() // parent of li
+			s.currentBlock = parentBlock
+			s.index++
+			return stateReadNextLiToken, nil
+		}
+		// ignore \n, if \n again, we must close current block
+		s.index++
+		return stateReadTextNewLine, nil
+	}
+	if char == '[' {
+		s.linkTitleValue = make([]byte, 0)
+		s.index++
+		return stateReadLinkTitle, nil
+	}
+	if char == '!' {
+		s.index++
+		return stateReadBeginImageToken, nil
+	}
+	if char == '`' {
+		s.currentBlock.Value = string(s.textValue)
+
+		parentBlock := s.blockStack.Top()
+		codeBlock := newBlock(TypeCode)
+		appendChild(parentBlock, codeBlock)
+		s.currentBlock = codeBlock
+
+		s.textValue = make([]byte, 0)
+		s.index++
+		return stateReadInlineCode, nil
+	}
+	// escape
+	if char == '<' {
+		s.textValue = appendStr(s.textValue, "&lt;")
+		s.index++
+		return stateReadText, nil
+	}
+	if char == '>' {
+		s.textValue = appendStr(s.textValue, "&gt;")
+		s.index++
+		return stateReadText, nil
+	}
+	if char == '&' {
+		s.textValue = appendStr(s.textValue, "&amp;")
+		s.index++
+		return stateReadText, nil
+	}
+	if char == '"' {
+		s.textValue = appendStr(s.textValue, "&quot;")
+		s.index++
+		return stateReadText, nil
+	}
+	s.textValue = append(s.textValue, char)
+	s.index++
+	return stateReadText, nil
 }
 
-type blockStack struct {
-	values []*Block
+func stateReadTextNewLine(s *parseState, char byte) (stateFunc, error) {
+	if char == '\n' {
+		// close all block
+		s.currentBlock.Value = string(s.textValue)
+		s.blockStack.Clear()
+		s.currentBlock = s.root
+		s.index++
+		return stateReadRootBlock, nil
+	}
+	s.textValue = append(s.textValue, char)
+	s.index++
+	return stateReadText, nil
 }
 
-func (s *blockStack) Push(v *Block) {
-	s.values = append(s.values, v)
+// link
+
+func stateReadLinkTitle(s *parseState, char byte) (stateFunc, error) {
+	if char == ']' {
+		s.index++
+		return stateReadLinkURLBeginToken, nil
+	}
+	s.linkTitleValue = append(s.linkTitleValue, char)
+	s.index++
+	return stateReadLinkTitle, nil
 }
 
-func (s *blockStack) Pop() *Block {
-	top := s.values[len(s.values)-1]
-	s.values = s.values[:len(s.values)-1]
-	return top
+// stateReadLinkURLBeginToken wants '('
+func stateReadLinkURLBeginToken(s *parseState, char byte) (stateFunc, error) {
+	if char == '(' {
+		s.linkURLValue = make([]byte, 0)
+		s.index++
+		return stateReadLinkURL, nil
+	}
+	s.textValue = append(s.textValue, '[')
+	s.textValue = appendStr(s.textValue, string(s.linkTitleValue))
+	s.textValue = append(s.textValue, ']')
+	// read current char as a part of text
+	return stateReadText, nil
 }
 
-func (s *blockStack) Top() *Block {
-	return s.values[len(s.values)-1]
+func stateReadLinkURL(s *parseState, char byte) (stateFunc, error) {
+	if char == ')' {
+		s.currentBlock.Value = string(s.textValue)
+
+		parentBlock := s.blockStack.Top()
+		linkBlock := newBlock(TypeAnchor)
+		linkBlock.Value = string(s.linkTitleValue)
+		linkBlock.URL = string(s.linkURLValue)
+		appendChild(parentBlock, linkBlock)
+
+		// next block
+		textBlock := newBlock(TypeText)
+		appendChild(parentBlock, textBlock)
+		s.currentBlock = textBlock
+
+		s.textValue = make([]byte, 0)
+		s.index++
+		return stateReadText, nil
+	}
+	s.linkURLValue = append(s.linkURLValue, char)
+	s.index++
+	return stateReadLinkURL, nil
 }
 
-func (s *blockStack) Clear() {
-	s.values = make([]*Block, 0)
+// image
+
+// stateReadBeginImageToken wants '['
+func stateReadBeginImageToken(s *parseState, char byte) (stateFunc, error) {
+	if char == '[' {
+		s.linkTitleValue = make([]byte, 0)
+		s.index++
+		return stateReadImageTitle, nil
+	}
+	s.textValue = append(s.textValue, '!')
+	// read this character as a part of text
+	return stateReadText, nil
+}
+
+func stateReadImageTitle(s *parseState, char byte) (stateFunc, error) {
+	if char == ']' {
+		s.index++
+		return stateReadImageURLBeginToken, nil
+	}
+	s.linkTitleValue = append(s.linkTitleValue, char)
+	s.index++
+	return stateReadImageTitle, nil
+}
+
+// stateReadImageURLBeginToken wants '('
+func stateReadImageURLBeginToken(s *parseState, char byte) (stateFunc, error) {
+	if char == '(' {
+		s.linkURLValue = make([]byte, 0)
+		s.index++
+		return stateReadImageURL, nil
+	}
+	s.textValue = append(s.textValue, '!')
+	s.textValue = append(s.textValue, '[')
+	s.textValue = appendStr(s.textValue, string(s.linkTitleValue))
+	s.textValue = append(s.textValue, ']')
+	// read current char as a part of text
+	return stateReadText, nil
+}
+
+func stateReadImageURL(s *parseState, char byte) (stateFunc, error) {
+	if char == ')' {
+		s.currentBlock.Value = string(s.textValue)
+
+		parentBlock := s.blockStack.Top()
+		imageBlock := newBlock(TypeImage)
+		imageBlock.Value = string(s.linkTitleValue)
+		imageBlock.URL = string(s.linkURLValue)
+		appendChild(parentBlock, imageBlock)
+
+		// next block
+		textBlock := newBlock(TypeText)
+		appendChild(parentBlock, textBlock)
+		s.currentBlock = textBlock
+
+		s.textValue = make([]byte, 0)
+		s.index++
+		return stateReadText, nil
+	}
+	if char == ' ' {
+		s.currentBlock.Value = string(s.textValue)
+
+		parentBlock := s.blockStack.Top()
+		imageBlock := newBlock(TypeImage)
+		imageBlock.Value = string(s.linkTitleValue)
+		imageBlock.URL = string(s.linkURLValue)
+		appendChild(parentBlock, imageBlock)
+
+		s.currentBlock = imageBlock
+
+		s.index++
+		return stateReadBeginImageAttr, nil
+	}
+	s.linkURLValue = append(s.linkURLValue, char)
+	s.index++
+	return stateReadImageURL, nil
+}
+
+func stateReadBeginImageAttr(s *parseState, char byte) (stateFunc, error) {
+	if char == ' ' {
+		s.index++
+		return stateReadBeginImageAttr, nil
+	}
+	if char == ')' {
+		parentBlock := s.blockStack.Top()
+		// next block
+		textBlock := newBlock(TypeText)
+		appendChild(parentBlock, textBlock)
+		s.currentBlock = textBlock
+
+		s.textValue = make([]byte, 0)
+		s.index++
+		return stateReadText, nil
+	}
+	s.attrName = make([]byte, 0)
+	return stateReadImageAttrName, nil
+}
+
+func stateReadImageAttrName(s *parseState, char byte) (stateFunc, error) {
+	if char == ' ' {
+		// empty value attr
+		attrName := string(s.attrName)
+		attrValue := ""
+		s.currentBlock.Attributes[attrName] = attrValue
+		s.index++
+		return stateReadBeginImageAttr, nil
+	}
+	if char == ')' {
+		// empty value attr
+		attrName := string(s.attrName)
+		attrValue := ""
+		s.currentBlock.Attributes[attrName] = attrValue
+
+		parentBlock := s.blockStack.Top()
+		// next block
+		textBlock := newBlock(TypeText)
+		appendChild(parentBlock, textBlock)
+		s.currentBlock = textBlock
+
+		s.textValue = make([]byte, 0)
+		s.index++
+		return stateReadText, nil
+	}
+	if char == '=' {
+		s.attrValue = make([]byte, 0)
+		s.index++
+		return stateReadImageAttrValue, nil
+	}
+	s.attrName = append(s.attrName, char)
+	s.index++
+	return stateReadImageAttrName, nil
+}
+
+func stateReadImageAttrValue(s *parseState, char byte) (stateFunc, error) {
+	if char == ' ' {
+		attrName := string(s.attrName)
+		attrValue := string(s.attrValue)
+		s.currentBlock.Attributes[attrName] = attrValue
+		s.index++
+		return stateReadBeginImageAttr, nil
+	}
+	if char == ')' {
+		attrName := string(s.attrName)
+		attrValue := string(s.attrValue)
+		s.currentBlock.Attributes[attrName] = attrValue
+
+		parentBlock := s.blockStack.Top()
+		// next block
+		textBlock := newBlock(TypeText)
+		appendChild(parentBlock, textBlock)
+		s.currentBlock = textBlock
+
+		s.textValue = make([]byte, 0)
+		s.index++
+		return stateReadText, nil
+	}
+	s.attrValue = append(s.attrValue, char)
+	s.index++
+	return stateReadImageAttrValue, nil
+}
+
+// UL
+
+func stateReadUL(s *parseState, char byte) (stateFunc, error) {
+	if char == ' ' {
+		if s.currentBlock.Type == TypeRoot {
+			// put ul
+			ulBlock := newBlock(TypeUL)
+			appendChild(s.currentBlock, ulBlock)
+			s.blockStack.Push(s.currentBlock)
+			s.currentBlock = ulBlock
+			s.index++
+			return stateReadFirstLiToken, nil
+		}
+		if s.currentBlock.Type == TypeUL {
+			s.index++
+			return stateReadFirstLiToken, nil
+		}
+	}
+	// just a paragraph
+	pBlock := newBlock(TypeP)
+	textBlock := newBlock(TypeText)
+	appendChild(s.currentBlock, pBlock)
+	appendChild(pBlock, textBlock)
+	s.blockStack.Push(s.currentBlock)
+	s.blockStack.Push(pBlock)
+
+	s.currentBlock = textBlock
+	s.textValue = make([]byte, 0)
+	s.textValue = append(s.textValue, '*')
+	// read this char as a part of text
+	return stateReadText, nil
+}
+
+func stateReadFirstLiToken(s *parseState, char byte) (stateFunc, error) {
+	if char == ' ' {
+		// skip
+		s.index++
+		return stateReadFirstLiToken, nil
+	}
+	// make li
+	liBlock := newBlock(TypeLI)
+	textBlock := newBlock(TypeText)
+	appendChild(s.currentBlock, liBlock)
+	appendChild(liBlock, textBlock)
+	s.blockStack.Push(s.currentBlock)
+	s.blockStack.Push(liBlock)
+
+	s.currentBlock = textBlock
+	s.textValue = make([]byte, 0)
+	// read this char as a part of text
+	return stateReadText, nil
+}
+
+func stateReadNextLiToken(s *parseState, char byte) (stateFunc, error) {
+	if char == ' ' {
+		// skip
+		s.index++
+		return stateReadNextLiToken, nil
+	}
+	if char == '*' || char == '-' {
+		s.index++
+		return stateReadFirstLiToken, nil
+	}
+	if char == '\n' {
+		// ul block is ended. clear all stack
+		s.blockStack.Clear()
+		s.currentBlock = s.root
+		s.index++
+		return stateReadRootBlock, nil
+	}
+	// new paragraph begins
+	s.blockStack.Clear()
+	s.currentBlock = s.root
+
+	pBlock := newBlock(TypeP)
+	textBlock := newBlock(TypeText)
+	appendChild(s.currentBlock, pBlock)
+	appendChild(pBlock, textBlock)
+	s.blockStack.Push(s.currentBlock)
+	s.blockStack.Push(pBlock)
+
+	s.currentBlock = textBlock
+	s.textValue = make([]byte, 0)
+	// read this char as a part of text
+	return stateReadText, nil
+}
+
+// pre code
+func stateReadBeginPreCode(s *parseState, char byte) (stateFunc, error) {
+	if char == '`' {
+		s.hCount++
+		if s.hCount < 3 {
+			s.index++
+			return stateReadBeginPreCode, nil
+		}
+		s.index++
+		return stateReadBeginPreCodeNewLine, nil
+	}
+	if s.hCount == 1 {
+		s.index++
+		return stateReadText, nil
+	}
+	s.textValue = make([]byte, 0)
+	return stateReadText, nil
+}
+
+func stateReadBeginPreCodeNewLine(s *parseState, char byte) (stateFunc, error) {
+	if char == '\n' {
+		// begin pre code
+		preCodeBlock := newBlock(TypePreCode)
+		textBlock := newBlock(TypeText)
+		appendChild(s.currentBlock, preCodeBlock)
+		appendChild(preCodeBlock, textBlock)
+		s.blockStack.Push(s.currentBlock)
+		s.blockStack.Push(preCodeBlock)
+		s.currentBlock = textBlock
+
+		s.textValue = make([]byte, 0)
+		s.index++
+		return stateReadPreCodeText, nil
+	}
+	// TODO support language
+	s.index++
+	return stateReadBeginPreCodeNewLine, nil
+}
+
+func stateReadPreCodeText(s *parseState, char byte) (stateFunc, error) {
+	if char == '`' {
+		s.hCount = 1
+		s.index++
+		return stateReadEndPreCode, nil
+	}
+	// escape
+	if char == '<' {
+		s.textValue = appendStr(s.textValue, "&lt;")
+		s.index++
+		return stateReadPreCodeText, nil
+	}
+	if char == '>' {
+		s.textValue = appendStr(s.textValue, "&gt;")
+		s.index++
+		return stateReadPreCodeText, nil
+	}
+	if char == '&' {
+		s.textValue = appendStr(s.textValue, "&amp;")
+		s.index++
+		return stateReadPreCodeText, nil
+	}
+	if char == '"' {
+		s.textValue = appendStr(s.textValue, "&quot;")
+		s.index++
+		return stateReadPreCodeText, nil
+	}
+	s.textValue = append(s.textValue, char)
+	s.index++
+	return stateReadPreCodeText, nil
+}
+
+func stateReadEndPreCode(s *parseState, char byte) (stateFunc, error) {
+	if char == '`' {
+		s.hCount++
+		if s.hCount < 3 {
+			s.index++
+			return stateReadEndPreCode, nil
+		}
+		// end of pre code
+		s.currentBlock.Value = string(s.textValue)
+
+		parentBlock := s.blockStack.Pop() // preCode
+		parentBlock = s.blockStack.Pop()  // parent of preCode
+		s.currentBlock = parentBlock
+		s.index++
+		return stateReadRootBlock, nil
+	}
+	for i := 0; i < s.hCount; i++ {
+		s.textValue = append(s.textValue, '`')
+	}
+	return stateReadPreCodeText, nil
+}
+
+// inline code
+
+func stateReadInlineCode(s *parseState, char byte) (stateFunc, error) {
+	if char == '`' {
+		s.currentBlock.Value = string(s.textValue)
+
+		textBlock := newBlock(TypeText)
+		parentBlock := s.blockStack.Top()
+		appendChild(parentBlock, textBlock)
+		s.currentBlock = textBlock
+
+		s.textValue = make([]byte, 0)
+		s.index++
+		return stateReadText, nil
+	}
+	s.textValue = append(s.textValue, char)
+	s.index++
+	return stateReadInlineCode, nil
+}
+
+func toHnType(level int) BlockType {
+	switch level {
+	case 1:
+		return TypeH1
+	case 2:
+		return TypeH2
+	default:
+		return TypeH1
+	}
 }
